Alerts: make severity consistant between sources and matching CAP protocol

+++ data/src/main/java/breezyweather/data/weather/WeatherRepository.kt
@@ -352,7 +352,7 @@ class WeatherRepository(
                     headline = alert.headline,
                     description = alert.description,
                     instruction = alert.instruction,
-                    severity = alert.severity.toLong(),
+                    severity = alert.severity,
                     color = alert.color.toLong()
                 )
             }
